v2

+++ Implementacion DAO y DTO/src/routes/admin.routes.js
@@ -1,8 +1,8 @@
-import express from "express";
+import { Router } from "express";
+import productModel from "../models/Products.js";
 import productDao from "../dao/productDao.js";
-import ProductDTO from "../dto/productDto.js";
 
-const adminRouter = express.Router();
+const adminRouter = Router();
 
 adminRouter.get("/", async (req, res) => {
   try {
@@ -11,24 +11,25 @@ adminRouter.get("/", async (req, res) => {
     const email = req.session.user.email;
     const rol = req.session.user.rol;
     const cartId = req.session.user.cartId;
-    const options = {
-      limit: parseInt(limit),
-      skip: (parseInt(page) - 1) * parseInt(limit),
-    };
-    console.log("Valor de cart:", cartId);
+    const options = {};
+    options.limit = parseInt(limit);
+    options.skip = (parseInt(page) - 1) * parseInt(limit);
 
     const queryOptions = query ? { title: { $regex: query, $options: "i" } } : {};
 
-    const totalCount = await productDao.countDocuments(queryOptions);
+    const totalCount = await productDao.getTotalProductCount(queryOptions);
     const totalPages = Math.ceil(totalCount / options.limit);
 
-    let products = await productDao.getProducts(queryOptions, options);
+    let productsQuery = productDao.getProducts(queryOptions, options);
 
+    // Verifica si se proporciona el parámetro 'sort' y aplica el método de ordenamiento ascendente y descendente
     if (sort === "asc") {
-      products = products.sort((a, b) => a.price - b.price);
+      productsQuery = productsQuery.sort({ price: 1 }); // Orden ascendente por precio
     } else if (sort === "desc") {
-      products = products.sort((a, b) => b.price - a.price);
+      productsQuery = productsQuery.sort({ price: -1 }); // Orden descendente por precio
     }
+
+    const products = await productsQuery;
 
     const response = {
       status: "success",
@@ -42,11 +43,10 @@ adminRouter.get("/", async (req, res) => {
       prevLink: page > 1 ? `http://localhost:4000/admin?limit=${limit}&page=${parseInt(page) - 1}` : null,
       nextLink: page < totalPages ? `http://localhost:4000/admin?limit=${limit}&page=${parseInt(page) + 1}` : null,
     };
-
-    res.render("admin", {
-      layout: false,
+    res.render('admin', {
+      layout: false, // Desactivar el uso del layout
       partials: {
-        navbar: "navbar",
+        navbar: 'navbar', // Incluir el partial del navbar si es necesario en otras vistas
       },
       products: products,
       response: response,
@@ -54,7 +54,7 @@ adminRouter.get("/", async (req, res) => {
       email: email,
       rol: rol,
       cartId: cartId,
-      message: message || "",
+      message: message || ""
     });
 
   } catch (error) {
@@ -62,14 +62,17 @@ adminRouter.get("/", async (req, res) => {
     res.status(500).send("Error al recibir los productos:");
   }
 });
-
+// Ruta para actualizar stock
 adminRouter.post("/products/:id/update-stock", async (req, res) => {
   try {
     const productId = req.params.id;
     const { amount } = req.body;
-
-    await productDao.updateStock(productId, parseInt(amount));
-
+    const product = await productDao.getProductById(productId);
+    if (!product) {
+      return res.status(404).send("Producto no encontrado");
+    }
+    product.stock = parseInt(amount);
+    await productDao.updateStock(productId, parseInt(amount)); 
     req.session.message = "Se ha actualizado el stock del producto.";
     res.redirect(`/admin?message=${encodeURIComponent(req.session.message)}`);
   } catch (error) {
@@ -78,13 +81,17 @@ adminRouter.post("/products/:id/update-stock", async (req, res) => {
   }
 });
 
+// Ruta para actualizar precio
 adminRouter.post("/products/:id/update-price", async (req, res) => {
   try {
     const productId = req.params.id;
     const { amount } = req.body;
-
-    await productDao.updatePrice(productId, parseInt(amount));
-
+    const product = await productDao.getProductById(productId);
+    if (!product) {
+      return res.status(404).send("Producto no encontrado");
+    }
+    product.price = parseInt(amount);
+    await productDao.updatePrice(productId, parseInt(amount)); // Utilizar el DAO para actualizar el precio
     req.session.message = "Se ha actualizado el precio del producto.";
     res.redirect(`/admin?message=${encodeURIComponent(req.session.message)}`);
   } catch (error) {
@@ -93,30 +100,42 @@ adminRouter.post("/products/:id/update-price", async (req, res) => {
   }
 });
 
-adminRouter.post("/products/:id/delete-product", async (req, res) => {
-  try {
-    const productId = req.params.id;
 
-    await productDao.deleteProduct(productId);
+  // Ruta para eliminar un producto del market
+  adminRouter.post("/products/:id/delete-product", async (req, res) => {
+    try {
+      const productId = req.params.id;
+      const product = await productDao.getProductById(productId);
+      if (!product) {
+        return res.status(404).send("Producto no encontrado");
+      }
+      await productDao.deleteProduct(productId);
+      req.session.message = "Has eliminado un producto.";    
+      res.redirect(`/admin?message=${encodeURIComponent(req.session.message)}`);
+    } catch (error) {
+      console.error("Error al eliminar un producto:", error);
+      res.status(500).send("Error al eliminar un producto");
+    }
+  });
 
-    req.session.message = "Has eliminado un producto.";
-    console.log(`Producto con ID ${productId} eliminado con éxito.`);
-  
-    res.redirect(`/admin?message=${encodeURIComponent(req.session.message)}`);
-  } catch (error) {
-    console.error("Error al eliminar un producto:", error);
-    res.status(500).send("Error al eliminar un producto");
-  }
-});
-
+// Ruta para agregar producto
 adminRouter.post("/products/owner/addproduct", async (req, res) => {
   try {
-    const { code, title, description, stock, id, price, thumbnail } = req.body;
+    const { code, title, description, stock, id, status, price, thumbnail } = req.body;
 
-    const productData = new ProductDTO(code, title, description, stock, id, price, thumbnail);
-
-    await productDao.addProduct(productData);
-
+    // Crear un nuevo producto utilizando el modelo y el email del usuario logeado
+    const newProduct = new productModel({
+      code,
+      title,
+      description,
+      stock,
+      id,
+      status: true,
+      price,
+      thumbnail,
+      owner: req.session.user.email, 
+    });
+    await productDao.addProduct(newProduct);
     req.session.message = "Producto agregado exitosamente.";
     res.redirect(`/admin?message=${encodeURIComponent(req.session.message)}`);
   } catch (error) {
